feat: Implement Loan Detail screen with Repayment Schedule and Documents display
-   Created `fragment_loan_detail.xml` to display detailed loan information.
-   Added `RepaymentScheduleAdapter` to display loan repayment schedule items in a `RecyclerView`.
-   Added `DocumentAdapter` to display loan documents in a `RecyclerView`.
-   Created `item_repayment_schedule.xml` layout for individual repayment schedule items.
-   Created `item_document.xml` layout for individual document items.
-   Updated `LoanDetailFragment` to:
    -   Receive and display loan details.
    -   Set up `RepaymentScheduleAdapter` and `DocumentAdapter`.
    -   Display borrower, collateral, and loan information.
    -   Handle empty documents state.
    -   Display repayment schedule and documents.
-   Updated `Loan` domain model to include `installments`.
-   Updated `InstallmentsItem` to include `Double` for `amountDue`.
-   Added base url to load images.
-   Added string resource for `documents`.
-   Updated base strings.

+++ core/src/main/java/com/ekachandra/core/data/source/remote/response/LoanResponse.kt
@@ -9,7 +9,7 @@ import com.google.gson.annotations.SerializedName
 data class InstallmentsItem(
 
     @field:SerializedName("amountDue")
-    val amountDue: Int? = null,
+    val amountDue: Double? = null,
 
     @field:SerializedName("dueDate")
     val dueDate: String? = null
